+++ plugin/evm/header/predicates.go
@@ -4,13 +4,8 @@
 package header
 
 import (
-	"errors"
-	"fmt"
-
 	"github.com/ava-labs/coreth/params"
 )
-
-var ErrInvalidExtraLength = errors.New("invalid header.Extra length")
 
 func ParsePredicates(rules params.AvalancheRules, extra []byte) ([]byte, error) {
 	if err := VerifyExtra(rules, extra); err != nil {
@@ -28,57 +23,4 @@ func ParsePredicates(rules params.AvalancheRules, extra []byte) ([]byte, error) 
 		return extra[offset:], nil
 	}
 	return nil, nil
-}
-
-// VerifyExtra verifies that the header's Extra field is correctly formatted for
-// [rules].
-func VerifyExtra(rules params.AvalancheRules, extra []byte) error {
-	extraLen := len(extra)
-	switch {
-	case rules.IsFUpgrade:
-		if extraLen < DynamicFeeAccumulatorSize {
-			return fmt.Errorf(
-				"%w: expected >= %d but got %d",
-				ErrInvalidExtraLength,
-				DynamicFeeAccumulatorSize,
-				extraLen,
-			)
-		}
-	case rules.IsDurango:
-		if extraLen < DynamicFeeWindowSize {
-			return fmt.Errorf(
-				"%w: expected >= %d but got %d",
-				ErrInvalidExtraLength,
-				DynamicFeeWindowSize,
-				extraLen,
-			)
-		}
-	case rules.IsApricotPhase3:
-		if extraLen != DynamicFeeWindowSize {
-			return fmt.Errorf(
-				"%w: expected %d but got %d",
-				ErrInvalidExtraLength,
-				DynamicFeeWindowSize,
-				extraLen,
-			)
-		}
-	case rules.IsApricotPhase1:
-		if extraLen != 0 {
-			return fmt.Errorf(
-				"%w: expected 0 but got %d",
-				ErrInvalidExtraLength,
-				extraLen,
-			)
-		}
-	default:
-		if uint64(extraLen) > params.MaximumExtraDataSize {
-			return fmt.Errorf(
-				"%w: expected <= %d but got %d",
-				ErrInvalidExtraLength,
-				params.MaximumExtraDataSize,
-				extraLen,
-			)
-		}
-	}
-	return nil
 }
